+++ custom_components/wemportal/number.py
@@ -8,9 +8,9 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 from . import get_wemportal_unique_id
-from homeassistant.helpers.entity import DeviceInfo
+from homeassistant.helpers.device_registry import DeviceInfo
 from .const import _LOGGER, DOMAIN
-from .utils import (fix_value_and_uom, uom_to_device_class)
+from .utils import (fix_value_and_uom, uom_to_device_class, build_device_info)
 
 
 async def async_setup_entry(
@@ -142,14 +142,7 @@ class WemPortalNumber(CoordinatorEntity, NumberEntity):
     @property
     def device_info(self) -> DeviceInfo:
         """Get device information."""
-        return {
-            "identifiers": {
-                (DOMAIN, f"{self._config_entry.entry_id}:{str(self._device_id)}")
-            },
-            "via_device": (DOMAIN, self._config_entry.entry_id),
-            "name": str(self._device_id),
-            "manufacturer": "Weishaupt",
-        }
+        return build_device_info(self._config_entry.entry_id, self._device_id)
 
     @property
     def available(self):